requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-pptx==0.6.21
reportlab==4.0.4
boto3==1.28.44
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
